Showing example of dynamic controller loading in winforms
This shows how layoutpanels can be used to add dynamic controller elements to a winforms application. The way we have discussed the design of th is project, we will need to parse a configuration file with the available options, and use the settings to build the form on the fly.