Added classes to represent properties of JSON config file
ProductLine - One of the options to select from the scroller on left side of UI, containing all information needed

IO - Class for the name of input/output variable and its type. For example, MeasuredHeight with a type float would belong to the Input property of one of the ProductLines

State - Representing a logical operation to be performed for a given ProductLine. Like dividing the height by 2 and then going to another state to evaluate available stock sizes.