Added Static UI elements, set up code for reading config files,
WIll be using Newtonsoft JSON package for deserializing JSON files. I added code to Program.cs to give an idea of how the files will need to be read. After the data is read, all the product lines will need to be added to the ProductLineSelector, and clicking on one should read the corresponding ProductLine object and build the UI based on the number of Inputs/Outputs and what type they are.