Changed JSON importing to use dynamic object to test state values for building correct state classes
State is an abstract class and can't be instantiated, therefore we treat the JSON as a dynamic and test the Operation property of each state to see what class it corresponds with. The example I added shows how we might build a DivisionState given what was read from the JSON object.