Added remaining static elements, set up code to read retrieve selected productline
Made changes necessary to the form to keep elements anchored to the screen and adjusted the look a bit, along with adding the table layouts needed for input and output

Also added code to retrieve the ProductLine object corresponding with the selected element. I could not expand on this code any further due to the input/output logic parameters being null, but will continue when this is fixed.