Added the state processing and messaging system
A new object StateMachine is used to represent each logic sequence (ie ResultingWidth) in the product line. Given input and a dictionary of booleans, it will process the input through each state until it encounters the end state, where it will return the resulting object. Debug.Write code is included in the parsing class; but will be moved to an official testing file down the line.