Reworked connection between UI and state output, and added wall jamb field
Major changes were required to add wall jambs to the ui since it is dependent on opening width. Therefore, I made in the product line config so now each output field is paired with the name of the input field it requires. Likewise, each state processing attempt is dependent on the output fields in the productline not the list of input fields. Product line generation code from program.cs were also moved into the productline constructor.